fix bug dupplication piste audio, delete page doublon. À faire : régler bug affichage page artist
